# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * BharatShield prototype risk analysis API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TextAnalysisInputLanguage = typeof TextAnalysisInputLanguage[keyof typeof TextAnalysisInputLanguage];


export const TextAnalysisInputLanguage = {
  en: 'en',
  hi: 'hi',
} as const;

export interface TextAnalysisInput {
  /**
     * @minLength 1
     * @maxLength 10000
     */
  text: string;
  language?: TextAnalysisInputLanguage;
}

export type FileAnalysisInputContentType = typeof FileAnalysisInputContentType[keyof typeof FileAnalysisInputContentType];


export const FileAnalysisInputContentType = {
  audio: 'audio',
  image: 'image',
  video: 'video',
} as const;

export interface FileAnalysisInput {
  /**
     * @minLength 1
     * @maxLength 255
     */
  fileName: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  fileType: string;
  /**
     * @minimum 1
     * @maximum 26214400
     */
  fileSize: number;
  contentType: FileAnalysisInputContentType;
  demo?: boolean;
}

export type AnalysisResultRiskLevel = typeof AnalysisResultRiskLevel[keyof typeof AnalysisResultRiskLevel];


export const AnalysisResultRiskLevel = {
  LOW: 'LOW',
  MEDIUM: 'MEDIUM',
  HIGH: 'HIGH',
  CRITICAL: 'CRITICAL',
} as const;

export type AnalysisResultContentType = typeof AnalysisResultContentType[keyof typeof AnalysisResultContentType];


export const AnalysisResultContentType = {
  text: 'text',
  audio: 'audio',
  image: 'image',
  video: 'video',
} as const;

export interface AnalysisResult {
  id: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  score: number;
  riskLevel: AnalysisResultRiskLevel;
  signals: string[];
  explanation: string;
  recommendations: string[];
  prototype: boolean;
  demo: boolean;
  contentType: AnalysisResultContentType;
  language: string;
  sourceLabel: string;
  createdAt: string;
}

export type HistoryItem = AnalysisResult;

export interface ModelStatus {
  name: string;
  status: string;
  description: string;
}

export interface ErrorResponse {
  error: string;
}