# lib/api-zod/src/generated/types/analysisResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * BharatShield prototype risk analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisResultContentType } from './analysisResultContentType';
import type { AnalysisResultRiskLevel } from './analysisResultRiskLevel';

export interface AnalysisResult {
  id: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  score: number;
  riskLevel: AnalysisResultRiskLevel;
  signals: string[];
  explanation: string;
  recommendations: string[];
  prototype: boolean;
  demo: boolean;
  contentType: AnalysisResultContentType;
  language: string;
  sourceLabel: string;
  createdAt: Date;
}