# lib/api-zod/src/generated/types/fileAnalysisInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * BharatShield prototype risk analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { FileAnalysisInputContentType } from './fileAnalysisInputContentType';

export interface FileAnalysisInput {
  /**
     * @minLength 1
     * @maxLength 255
     */
  fileName: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  fileType: string;
  /**
     * @minimum 1
     * @maximum 26214400
     */
  fileSize: number;
  contentType: FileAnalysisInputContentType;
  demo?: boolean;
}