# lib/api-zod/src/generated/types/historyItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * BharatShield prototype risk analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisResult } from './analysisResult';

export type HistoryItem = AnalysisResult;