# lib/api-zod/src/generated/types/modelStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * BharatShield prototype risk analysis API
 * OpenAPI spec version: 0.1.0
 */

export interface ModelStatus {
  name: string;
  status: string;
  description: string;
}