# lib/api-zod/src/generated/types/textAnalysisInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * BharatShield prototype risk analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { TextAnalysisInputLanguage } from './textAnalysisInputLanguage';

export interface TextAnalysisInput {
  /**
     * @minLength 1
     * @maxLength 10000
     */
  text: string;
  language?: TextAnalysisInputLanguage;
}